Rename shared-utils -> common
