+++ mocks/utils.ts
@@ -1,4 +1,4 @@
-import { BASE_PATH } from '../src/shared-utils/miljø';
+import { BASE_PATH } from '../src/common/miljø';
 
 export function urlMedBasePath(path) {
     return BASE_PATH + path;
