+++ src/backend/middlewares/modell-versjon-interceptor.test.ts
@@ -3,7 +3,7 @@ import { mockDeep } from 'vitest-mock-extended';
 
 import { type ApiRessurs, RessursStatus } from '@navikt/familie-typer';
 
-import { modellMismatchMelding, ModellMismatchRespons, modellVersjon } from '../../shared-utils/modellversjon';
+import { modellMismatchMelding, ModellMismatchRespons, modellVersjon } from '../../common/modellversjon';
 
 import { modellVersjonInterceptor } from './modell-versjon-interceptor';
