+++ src/backend/middlewares/proxy.ts
@@ -7,7 +7,7 @@ import { v4 as uuid } from 'uuid';
 
 import { logError, logSecure } from '@navikt/familie-logging';
 
-import { erLokalt } from '../../shared-utils/miljø';
+import { erLokalt } from '../../common/miljø';
 
 import { AUTHORIZATION_HEADER } from './tokenProxy';
