+++ src/backend/routes/feature-toggles.ts
@@ -4,7 +4,7 @@ import { ParamsDictionary } from 'express-serve-static-core';
 import { byggSuksessRessurs, type Ressurs } from '@navikt/familie-typer';
 
 import { EAllFeatureToggles, defaultFeatureToggleValues, ToggleKeys } from '../../frontend/typer/feature-toggles';
-import { BASE_PATH } from '../../shared-utils/miljø';
+import { BASE_PATH } from '../../common/miljø';
 import { isEnabled } from '../utils/unleash';
 
 const fetchAllFeatureTogglesHandler: RequestHandler<ParamsDictionary, Ressurs<EAllFeatureToggles>> = (_, res) => {
