+++ src/backend/routes/modellversjon.ts
@@ -2,8 +2,8 @@ import { Express } from 'express';
 
 import { byggSuksessRessurs } from '@navikt/familie-typer';
 
-import { BASE_PATH } from '../../shared-utils/miljø';
-import { modellVersjon } from '../../shared-utils/modellversjon';
+import { BASE_PATH } from '../../common/miljø';
+import { modellVersjon } from '../../common/modellversjon';
 
 export const konfigurerModellVersjonEndpoint = (app: Express): Express => {
     app.get(`${BASE_PATH}modellversjon`, (_, res) => {
