+++ src/common/miljø.ts
