+++ src/frontend/App.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter as Router } from 'react-router';
 
 import { GlobalAlert } from '@navikt/ds-react';
 
-import { BASE_PATH } from '../shared-utils/miljø';
+import { BASE_PATH } from '../common/miljø';
 
 import AppContainer from './AppContainer';
 import { AppProvider } from './context/AppContext';
