+++ src/frontend/components/Felleskomponenter/AppLenke/AppLenke.tsx
@@ -4,8 +4,8 @@ import { useNavigate } from 'react-router';
 
 import { Link } from '@navikt/ds-react';
 
-import { BASE_PATH } from '../../../../shared-utils/miljø';
-import { unslash } from '../../../../shared-utils/unslash';
+import { BASE_PATH } from '../../../../common/miljø';
+import { unslash } from '../../../../common/unslash';
 import { useAppNavigationContext } from '../../../context/AppNavigationContext';
 import { ISteg } from '../../../typer/routes';
