+++ src/frontend/components/SøknadsSteg/Oppsummering/Oppsummeringsbolk.tsx
@@ -5,8 +5,8 @@ import { useNavigate } from 'react-router';
 import { FormSummary } from '@navikt/ds-react';
 import type { ISkjema } from '@navikt/familie-skjema';
 
-import { BASE_PATH } from '../../../../shared-utils/miljø';
-import { unslash } from '../../../../shared-utils/unslash';
+import { BASE_PATH } from '../../../../common/miljø';
+import { unslash } from '../../../../common/unslash';
 import { useAppContext } from '../../../context/AppContext';
 import { useStegContext } from '../../../context/StegContext';
 import { IBarnMedISøknad } from '../../../typer/barn';
