+++ src/frontend/context/FeatureTogglesContext.tsx
@@ -2,7 +2,7 @@ import React, { createContext, PropsWithChildren, useContext, useState } from 'r
 
 import { type Ressurs, RessursStatus } from '@navikt/familie-typer';
 
-import { BASE_PATH } from '../../shared-utils/miljø';
+import { BASE_PATH } from '../../common/miljø';
 import useFørsteRender from '../hooks/useFørsteRender';
 import { EAllFeatureToggles, defaultFeatureToggleValues } from '../typer/feature-toggles';
