+++ src/frontend/context/SanityContext.tsx
@@ -4,7 +4,7 @@ import { createClient } from '@sanity/client';
 
 import { byggHenterRessurs, byggTomRessurs, type Ressurs, RessursStatus } from '@navikt/familie-typer';
 
-import miljø from '../../shared-utils/miljø';
+import miljø from '../../common/miljø';
 import { SanityDokument } from '../typer/sanity/sanity';
 import { ITekstinnhold } from '../typer/sanity/tekstInnhold';
 import { transformerTilTekstinnhold } from '../utils/sanity';
