+++ src/frontend/context/pdl.ts
@@ -1,6 +1,6 @@
 import type { Ressurs } from '@navikt/familie-typer';
 
-import miljø from '../../shared-utils/miljø';
+import miljø from '../../common/miljø';
 import { ISøkerRespons } from '../typer/person';
 
 export const hentSluttbrukerFraPdl = (axiosRequest): Promise<Ressurs<ISøkerRespons>> => {
