+++ src/frontend/decorator.ts
@@ -1,6 +1,6 @@
 import { injectDecoratorClientSide } from '@navikt/nav-dekoratoren-moduler';
 
-import { erProd } from '../shared-utils/miljø';
+import { erProd } from '../common/miljø';
 
 export function hentDekorator() {
     const env = erProd() ? 'prod' : 'dev';
