+++ src/frontend/hooks/useSendInnSkjema.tsx
@@ -3,8 +3,8 @@ import { AxiosError } from 'axios';
 
 import { type Ressurs, RessursStatus } from '@navikt/familie-typer';
 
-import miljø from '../../shared-utils/miljø';
-import { erModellMismatchResponsRessurs } from '../../shared-utils/modellversjon';
+import miljø from '../../common/miljø';
+import { erModellMismatchResponsRessurs } from '../../common/modellversjon';
 import { useAppContext } from '../context/AppContext';
 import { useSpråkContext } from '../context/SpråkContext';
 import { ISøknadKontrakt } from '../typer/kontrakt/søknadKontrakt';
