+++ src/webpack/webpack.production.config.ts
@@ -5,7 +5,7 @@ import TerserWebpackPlugin from 'terser-webpack-plugin';
 import webpack from 'webpack';
 import { CustomizeRule, mergeWithRules } from 'webpack-merge';
 
-import { BASE_PATH } from '../shared-utils/miljø';
+import { BASE_PATH } from '../common/miljø';
 
 import baseConfig from './webpack.common.config';
